+++ s3/src/test/scala/akka/stream/alpakka/s3/impl/SplitAfterSizeSpec.scala
@@ -127,6 +127,6 @@ class SplitAfterSizeSpec(_system: ActorSystem)
     )
   }
 
-  def bytes(start: Byte, end: Byte): Array[Byte] = (start to end).map(_.toByte).toArray[Byte]
+  def bytes(start: Byte, end: Byte): Array[Byte] = (start.toInt to end.toInt).map(_.toByte).toArray[Byte]
 
 }
